Mesa; GLU: Mention libglvnd.
This commit specifically adds notes for both Mesa and GLU, and the notes
are different for each package.

For Mesa: the note mentions that Mesa provides OpenGL libraries, and
compatibility is lower with packages and binaries outside of BLFS. Thus,
the user can install libglvnd instead for OpenGL libraries. I mentioned
that this is not supported by the BLFS editors.

For GLU: the note mentions that if libglvnd is installed, the user must
install the GLFS version of GLU instead. The differences between the
BLFS and GLFS versions are very small. GLU by default requests opengl.pc
which is not provided by Mesa, but is provided by libglvnd. When it
can't find opengl.pc, configuration fails. BLFS solves this by forcing
meson to use gl.pc instead. On a libglvnd system, GLU using gl.pc makes
libGLU function similarly to a noop function. GLFS solves this behavior
by doing:
{{{
$(! pkgconf libglvnd && echo -D gl_provider=gl).
}}}

It might be better to replace `libglvnd` with `opengl` to be more
direct, as well as removing the `!` and replacing `&&` with `||`. Either
way, the statement checks to see if libglvnd can be found. If it can't,
it tells meson that only gl.pc is present. It could be easily adapted
for BLFS, but it just makes more sense for that compatibility to remain
in GLFS at the moment.